don't build targets in `comp` in parallel
Otherwise, src/comp/Makefile invokes multiple ghc builds, all in the
same workspace, which break each other sometimes.

We still get some parallelization, by passing -jN to GHC itself, but
reading what number of jobs is passed to make, and passing it down to
GHC doesn't really work
[without a lot of hackery](https://stackoverflow.com/questions/5303553),
and there seems to be a
[upper limit](https://gitlab.haskell.org/ghc/ghc/issues/9221) as well,
so let's pick a "reasonable number", and allow overriding it via
GHCJOBS=n, if people run it on a very beefy machine.

Fixes #38